// template/src/Components/Core/DataDisplay/Typography/type.ts
export enum TYPOGRAPHY_VARIANT {
  /**
   * 20px, font-weight 400, line-height 28px
   */
  REGULAR_20 = 'REGULAR_20',
  /**
   * 18px, font-weight 400, line-height 26px
   */
  REGULAR_18 = 'REGULAR_18',
  /**
   * 16px, font-weight 400, line-height 26px
   */
  REGULAR_16 = 'REGULAR_16',
  /**
   * 15px, font-weight 400, line-height 20px
   */
  REGULAR_15 = 'REGULAR_15',
  /**
   * 14px, font-weight 400, line-height 20px
   */
  REGULAR_14 = 'REGULAR_14',
  /**
   * 13px, font-weight 400, line-height 20px
   */
  REGULAR_13 = 'REGULAR_13',
  /**
   * 13px, font-weight 400, line-height 18px
   */
  REGULAR_12 = 'REGULAR_12',
  /**
   * 11px, font-weight 400, line-height 16px
   */
  REGULAR_11 = 'REGULAR_11',
  /**
   * 20px, font-weight 500, line-height 28px
   */
  MEDIUM_20 = 'MEDIUM_20',
  /**
   * 18px, font-weight 500, line-height 26px
   */
  MEDIUM_18 = 'MEDIUM_18',
  /**
   * 16px, font-weight 500, line-height 26px
   */
  MEDIUM_16 = 'MEDIUM_16',
  /**
   * 15px, font-weight 500, line-height 20px
   */
  MEDIUM_15 = 'MEDIUM_15',
  /**
   * 14px, font-weight 500, line-height 20px
   */
  MEDIUM_14 = 'MEDIUM_14',
  /**
   * 13px, font-weight 500, line-height 20px
   */
  MEDIUM_13 = 'MEDIUM_13',
  /**
   * 12px, font-weight 500, line-height 18px
   */
  MEDIUM_12 = 'MEDIUM_12',
  /**
   * 11px, font-weight 500, line-height 16px
   */
  MEDIUM_11 = 'MEDIUM_11',
  /**
   * 44px, font-weight 700, line-height 48px
   */
  BOLD_44 = 'BOLD_44',
  /**
   * 40px, font-weight 700, line-height 48px
   */
  BOLD_40 = 'BOLD_40',
  /**
   * 28px, font-weight 700, line-height 36px
   */
  BOLD_28 = 'BOLD_28',
  /**
   * 24px, font-weight 700, line-height 30px
   */
  BOLD_24 = 'BOLD_24',
  /**
   * 22px, font-weight 700, line-height 30px
   */
  BOLD_22 = 'BOLD_22',
  /**
   * 20px, font-weight 700, line-height 28px
   */
  BOLD_20 = 'BOLD_20',
  /**
   * 18px, font-weight 700, line-height 26px
   */
  BOLD_18 = 'BOLD_18',
  /**
   * 16px, font-weight 700, line-height 26px
   */
  BOLD_16 = 'BOLD_16',
  /**
   * 15px, font-weight 700, line-height 20px
   */
  BOLD_15 = 'BOLD_15',
  /**
   * 14px, font-weight 700, line-height 20px
   */
  BOLD_14 = 'BOLD_14',
  /**
   * 13px, font-weight 700, line-height 20px
   */
  BOLD_13 = 'BOLD_13',
  /**
   * 12px, font-weight 700, line-height 18px
   */
  BOLD_12 = 'BOLD_12',
  /**
   * 11px, font-weight 700, line-height 16px
   */
  BOLD_11 = 'BOLD_11',
  /**
   * 10px, font-weight 700, line-height 12px
   */
  BOLD_10 = 'BOLD_10',
}

export type TypographyVariantLiteral = keyof typeof TYPOGRAPHY_VARIANT;